Move sync.js to lib folder
