+++ lib/sync.js
@@ -5,11 +5,11 @@ var async = require('async');
 var rm = require('rimraf');
 var debug = require('debug')('sync:sync');
 
-var git = require('./lib/git');
-var trash = require('./lib/trash');
-var parse = require('./lib/parse');
-var oss = require('./lib/oss');
-var progressBar = require('./lib/progress');
+var git = require('./git');
+var trash = require('./trash');
+var parse = require('./parse');
+var oss = require('./oss');
+var progressBar = require('./progress');
 
 exports = module.exports = Sync;
