+++ package.json
@@ -2,7 +2,7 @@
   "name": "oss-sync",
   "version": "0.1.1",
   "description": "Aliyun OSS sync tool",
-  "main": "sync.js",
+  "main": "index.js",
   "author": "mehwww",
   "license": "MIT",
   "repository": {
